(#5) Removed hard coded file path
This has now been promoted to a ToolSettings property, which can be
overridden if/when required. It still defaults to the previously
hard-coded location, but it will first attempt to locate the tool using
the Tools.Resolve method, and it that can't be found, it will register
a tool at the previously hard-coded location.