(#10) Add support for build/sign of MSI
This is based on the assumption that there is a single solution file
that generates the MSI's, using a configuration value of "WIX".

It is possible to control what MSI's are signed based on a parameter
similar to what is done with signing of PowerShell Scripts and
Assemblies.

By default, no building of MSI's will happen, and you will need to set
the shouldBuildMsi parameter to true in the recipe.cake file to get
things to build as expected.