(#15) Use package version as TeamCity build number
When running on TeamCity, send a service message to update the TeamCity
build number to be the same as the asserted package version. Prior to
this, the version number that GitVersion asserted would be getting sent
to TeamCity, which is not exactly the same as the version number that
we are using for our packages.