(#16) Additional check for running tasks
We only want to run the signing steps when we _know_ that the required
files are in place on the machine where the build is running.
Historically, this was done using the presence of the certificate on
disk.  However, when running on TeamCity, this isn't the case, and
insead the certificate is sourced from a Ceritificate store.

Here the criterias have been updated to check to see whether build is
runnng on Teamcity, OR, the certiificate is located in the right place.