(#24) Correctly assert branch name
TeamCity uses the complete name of a branch, for example
refs/heads/develop, however, the Recipe uses the short (simple) name of
the branch.  We need to ensure that this is returned from the build
provider.

In addition, we need to handle the situtation where the branch being
built is a tag, and find the underlying branch that the tag is
associated with. A compromise is made here, taking the first associated
branch for the tag.