(#27) Remove usage of BuildVersion setup context
This was brought over from the Cake.Recipe project, however, this
hasn't been setup in this Recipe, yet.  We still have everything in the
BuildParameters class.

At some point, we are going to want to switch to using this, but not
just now.  This didn't cause any problems when running locally, only
when running on TeamCity.  This was due to the task being skipped
locally, and the setup context never having to be inspected.