(#29) Ensure that Milestone is set to a value
Rather than null, which is what is was doing before.  In addition, make
sure to output all the asserted values, as well as to be consistent
about the names and ordering of the properties within the file.

Finally, if the build is being run on a copy of the repository, that
isn't a clone of the repository, provide some fake version numbers, and
warn of the problem.