(#30) Don't duplicate pushing of packages
Make the pushing of pre-release and release packages mutually
exclusive, to prevent the same package being pushed to the same feed
(when they are configured to be the same) thus causing an error.

In addition, add a criteria to both the creation of Chocolatey and
NuGet packages to allow preventing of even attempting to run these
tasks when switched off at the highest level.  This is similar to what
is done with the ShouldRunDotNetCorePack parameter.