(#33) Truncate prerelease label
If the prerelease label gets too long, it can cause problems when
running things like choco pack, since the package version is too long.
This typically happens when running on a local feature branch, or as
part of a PR.

To "fix" this, for now, simply trancate the prerelease label to a
maximum of 10, which will allow things to continue. It is likely that
more smarts will be needed here, at some point, but for now, this is
fine.