(#35) Be consistent about skipped tasks
If we are going to skip one unit testing task if we are running an
official build, we should be consistent and skip all the related
tasks.

At some point, we might want to re-think the concept of skipping these
tasks, but for now, consistency is better than anything else.