(#34) Add support for ILMerge'ing
This is achieved using the ilmerge NuGet package:

https://www.nuget.org/packages/ilmerge/

It is possible to configure what tool is used (i.e. what version) via
the normal ToolSettings.SetToolPreprocessorDirectives method.

It will be the responsibility of the consuming project to define
exactly what files need to be ILMerged, including what assembly,
keyfile, etc, are used.

It looks like the Cake alias for ILMerge isn't quite right, as the
internalize parameter isn't a boolean, and it is missing parameters
like keyfile.  Will need to do a PR into Cake at some point to correct
this.