(#36) Remove replacement of forward slash
When adding a path to the assemblyinfo.cs file, the backslashes need to
be escaped, otherwise, it causes errors. However, there was also a
replacement of valid forward slashes (which are used on POSIX systems).
This problem only showed up once we started building the
chocolatey/choco repository on GitHub Actions across multiple different
operating systems.