(#40) Update to Cake.Git 1.1.0
This is to prevent the error that is being seen on the GitHub Actions
Ubuntu image, due to the bundled version of the LibGit2Sharp
assembly.

We can't upgrade to the latest latest version, since this doesn't work
with the version of Cake that we are currently using in the Recipe.