(#43) Allow forcing usage of .NET Global Tools
Rather than solely deciding which version of the tools to use based on
what type of project is being built, also allow the decision to be forced
using a build parameter.

This is necessary for example when running GitVersion on the GitHub
Actions Ubuntu image, where the .NET Framework version doesn't work
at all.