(#52) Update to latest GitVersion package
We were previously pinning to a specific version of the package because
it was known that this was the last version that worked under Mono when
running on Posix.  However, this is no longer the case, since this
package version no longer worked on the latest Ubuntu images.

In these instances, the recommendation is to use the .NET Global Tool
version of GitVersion, which can be done via a parameter in the
recipe.cake file.