(#55) Add support for generating lcov report
OpenCover, which is used for gneerating a test coverage report
generates an XML file with the results. This format of report can't be
published to coveralls.io since the tool that we are using doesn't
support it.

Use ReportGenerator (had to also update to a newer version for this
support) to convert the xml file to a lcov file, which we can then
report to coveralls.io.