(#56) Update all artifacts through build provider
This will mean that they don't have to be explicitly defined in the CI
configuration files.

Uploading artifacts to GitHub Actions will still need to be defined,
since the version of Cake that we are using doesn't support uploading
to GitHub Actions, but once we update to the latest Cake, things will
just start working.