(#23) Add support for integration tests
Introduce a new argument, testExecutionType, which can be set to either
unit, integration, or all, to control which test assemblies are exectued.

This controls the TestAssemblyFilePattern, which sets the globbing
pattern used to locate the test assemblies.

The default for this will be unit tests, but when required, this can be
set higher to execute additional tests.