(#64) Allow different usage of EazFusactor
When ILMerge'ing and obfuscating assemblies (so far, only on full fat
.NET Framework projects) it is necessary to pass all the files at once
into EazFuscator, rather than one at a time.

Passing them one at a time was found to be necessary when working with
dotnet projects, since it was necessary to also pass in the MSBuild
project file to get things to work.

Since there seems to be two ways that this needs to operate (at least
for now) switch how EazFuscator is called based on whether it is a full
fat .NET Framework project, or the newer format.