(maint) Tidy up of parameters/toolsettings files
Took the opportunity to do a general formatting of these two files. This
included making sure that all properties, parameters, etc were in
alphabetical order, making them easier to find.  In addition, made sure
to assert if required input variables are not passed in, and throw an
error if one isn't provided.