(#65) Remove usage of NuGet.config file
The configuration of sources is now done exclusively through the
default values, augmented by the NUGETDEVRESTORE_SOURCE environment
variable, or through the PRIMAY_NUGET_SOURCE environment variable, or
through the usage of the nuGetSources parameter.

It was found that an "empty" NuGet.config file at the root of the
repository to control some other NuGet.exe settings was cuasing a
problem with the restoring of packages. Multiple attempts were made to
work around this, but none had the desired affect.  As a result, it was
decided to completely remove the usage of the NuGet.config file.