(#66) Add shouldGenerateSolutionVersionCSharpFile
This parameter is used to control whether the SolutionVersion.cs file
is generated or not. This defaults to true, as most projects still
require this file, however, there are starting to be some projects
where this file is not required.

At some point, this parameter may flip to being off by default.