(#71) Add handling for shouldRunOpenCover argument
In addition to being able to control running of OpenCover from a
parameter in the recipe.cake file, this commit allows this to be
controlled from the command line also, using the --shouldRunOpenCover
argument here.  This argument will override anything that is set in the
recipe.cake file.

The need for this change came from the requirement to be able to run
OpenCover for some test executions, but not others.  i.e. running the
Integration Tests for chocolatey/choco under OpenCover isn't working,
although not entirely clear why, but running OpenCover for the unit
tests is working fine.

This command line argument will mean that we can control when to use
OpenCover between build configuration, without the need to modify the
code in the repository.