(#72) Remove modification of strong name keys
This commit removes the stripping of the strong name key from a
Chocolatey assembly when it is referenced by a project.  This was
previously done in an attempt to ensure that the referenced Chocolatey
assemblies had the "correct" strong name key for the type of build that
was being done, i.e. official or non-official.  During initial testing,
this appeared like it was working correctly, but once we got to the
Chocolatey Agent build, things started to fall-over.  This task is now
being removed, and we will revert back to the previous way of doing
this, using the Get-ChocoUpdatedDebugVersion.ps1 scripts that exist in
the required repositories.