(#80) Add support for publishing AWS Lambdas
By default, projects which are identified as AWS Lambda projects, will
be published as a zip file to a new _PublishedLambdas folder in the
code_drop folder.  These zip files will also be published as artifacts
in the CI Server.

Later, in the build process, assuming that there is a tag, this zip
file will then be published to AWS (using the AWS environment variables
which are assumed to be present on the system).  Currently, there are
no checks that these exist, since this could be done via a profile file
as well.