(#80) Remove need to globally install Lambda tool
Instead, the Amazon.Lambda.Tools .NET Global Tool package is installed
locally to the project, and executed directly from there.  This is why
it has switched from DotNetCoreTool to StartProcess.

Ideally, at some point, a Cake Addin will be created to allow direct
execution of this exe, however, for now, we will stick with using
StartProcess.