(#82) Remove version from function name if present
We ran into an issue with publishing of the AWS Lambda, if a version
number is present in the zip file name. This would mean that the publish
can't happen, since the function name would not match exactly, what as
present in AWS.

The removal of the version number is done by using one of the matched
groups that comes from the regular expression, which only contains the
function name.