(#86) Add XBuildPlatformTarget to ToolSettings
This will allow the control of what buid platform target is used when
calling XBuild.  An equivalent setting currently exists when calling
MSBuild, but not for XBuild, so this brings consistency in this area.

It was found that this setting is required, since when running
./build.official.sh on a POSIX machine, the output folder for the
chocolatey.console project was not the same as when running on a
Windows Machine.