(#87) Fix dependencies for AWS Lambda task
Only run DotNet Tasks, when the RunDotNet method has been selected
within the recipe.cake file. This will prevent any other task running
these tasks via a dependency introduced to the recipe.

Additionally, only hook up the dependency between the
PublishAwsLambdasTask and DotNetBuild, when it is known that we are
running a DotNet Build.  This "cleans" up which tasks are executed to
only those that _need_ to be executed.