(#85) Add globbing pattern for integration tests
Previously, there was an assumption that a single globbing pattern that
could be provided that would allow the selection of the files that
needed to be tested, for both unit and integration tests.  However, in
practice, this turned out to not be teh case, as it was
over-somplicated, and prone to error.

This commit switches to having two globbing patterns, one for unit
tests and another for integration tests, and based on the test
execution type that is running, these globbing patterns will either be
used individually, or combined.