(#89) Stop null reference exception on branchName
There are times, for example, when you have checked out a tag directly,
that there is no branchName in play (from a git perspective).  In these
scenarios, when attempting to assert the BranchType, we need to handle
gracefully this situation, so that it doesn't cause a null reference
exception.

Add a simple null check when the branchType is being asserted, which
will then make it flow through to BranchType.Unknown.