(#90) Add configurable globbing patterns
For both NuGet and Chocolatey nuspec and nupkg files. This will allow
finer grained control over exactly which nuspec files are selected to
be packed, and also which of the generated nupkgs need to be pushed to
a repository.

The need for this was found with the chocolatey/choco build, where we
now create a no7zip version of the package, which we don't want/need to
be published to a repository, but which we do want to be created, and
reside within the same folder structure as the other artifacts.