(#91) Defer throwing exceptions on test failures
This commit updates the tasks for running unit/integration tests
to add an OnError handler so the actual throwing of the exception
can be defered to after we have uploaded the test results to any
build servers, or when generating local coverage files.

Additionally the unit and coverage results generation continues
with any additional tasks if one of these fails.