(#93) Add support for resolving VS with VSWhere
This commit adds new handling in the ToolsSettings class to
resolve to the actual instance of an AMD64 MSBuild executable
to ensure we are using either the supported VS studio instance
or the latest instance available.

This change is made due to the version of cake that we are
currently using isn't able to resolve to a Visual Studio 2022
instance, or if an instance is installed to a custom directory.