(#105) Allow local build to upload artifacts
When building locally, there is no way of knowing what files would be
uploaded for publishing on Team City or GitHub Actions other than by
reading the command output to find everywhere that we indicate it would
have uploaded. This commit fixes that by creating a new Artifacts
directory in the code_drop directory and copying any file asked to be
uploaded into that directory.