(#107) Add sending of notifications
This commit adds the ability to send notifications to Mastodon, Discord,
and Slack. There is also a partial implementation of posting to Twitter
but this is known not to work, since there has been a change to the
Twitter API.  In order for notifications to be sent, the necessary
environment variables need to be set.  By default, an attempt will be
made to send a notification to all configured destinations, when running
--target=send-notifications, but this can be overridden either through
setting the "should" parameter in the recipe.cake file, or by passing
the "should" command line option to false.