(#112) Make Build-MSI task only run on Windows
In order to build an MSI, we need to use WiX, and currently, that only
works when running the build on a Windows machine, so we need to be
more explicit about this.  We currently run the build for
chocolatey/choco on Windows, Mac, and Linux, and the tagged builds fail
when attempting to build the MSI on Mac and Linux.  As a result, the
publishing of the Docker images fail, since those builds run on Linux.