(#114) Update to latest Cake.Twitter addin
And also re-enable the sending of Tweets, rather than just writing it
to the output.

The latest version of the Cake.Twitter addin has enabled support for
the new Twitter v2 API, since the old API no longer functions.  This
means that automating of tweets to Twitter is now ready for usage.