(#117) Add ShouldRunTests option/parameter
A new ShouldRunTests option and parameter have been added to the
recipe script, which will allow the bypassing of all tasks associated
with testing of the code base.  This will be possible through a
parameter in the recipe.cake file, as well as using a command line
option to set this to false.

This functionality is needed to make it possible to speed up builds
running locally on your machine, as well as tailoring CI builds if/when
required.