(#118) Add CLI options for all "should" parameters
This will give the person running the build, either manually at the CLI
or through CI/CD, direct control over exactly which tasks are run,
without the need to make direct changes to the source control files to
make it happen.

The command line option will take priority over what is configured in
the recipe.cake file, so it will be possible to override what has been
configured there, if/when required.