(#119) Change dependencies are run time
Rather than taking a hard dependency on the "Build" task, make it so
that this is defined based on whether a .NET Core build, or a full fat
.NET Framework build is running.  This will mean that the SonarQube
initialization and finalization will complete at the correct time,
rather than running more tasks than it was meant to, i.e. running the
Build and Restore tasks when it shouldn't.

This was missed in the initial creation of the SonarQube tasks, as this
was only run against Chocolatey CLI, and not some of the other
projects, that use .NET Core.