(#120) Add ability to run dotnet-format
This will be executed as part as the AnalyzeTask, and by default, no
changes will be made to any source files, i.e. run the --check option.
A report will be generated in the TestResults folder.

An additional Task, Run-DotNetFormat, which can be executed to actually
perform the fixups that were found.

There is a top level parameter, shouldRunDotNetFormat, which can be
used to control whether this check task runs.  This can be set within
the recipe.cake file, and as a command line option.  This isn't set on
the Run-DotNetFormat task, since this task has to be run directly,
therefore you are explicitly opting in to run that task, so no
additional criteria is required.