(#110) Be explicit about when tasks should run
Both the SonarQube and DependencyCheck tasks should only be run when
running in TeamCity, when it is on the develop branch, and when not a
pull request. This is the only time that all the required environment
variables and infrastructure is in place for it to complete.

It is not envisioned that these tasks will ever have to be run locally,
so the control for passing in shouldRunSonarQube and
shouldRunDependencyCheck is not required, neither at the command line,
or via the recipe.cake.  The ability to stop these tasks from doing
anything can be controlled via the environment variable for SonarQube
itself, if required.