(#122) Fix dependency graph for unit testing
When running a project via .NET Core, the dependency graph wasn't being
setup properly when it comes to unit testing. The end results was that
tasks were running to act on the results of the unit tests, when the
unit tests hadn't actually been run yet.

This commit address this problem by ensuring that all dependencies are
in place, similar to how they are done when it is being run under full
.NET Framework.