(#124) Remove usage of IsNetCore method
This method is known to cause problems, as it doesn't accurately detect
whether a project is .NET Core on anything above .NET 5. To that end,
let's simply remove the usage of this method, and proceed with the usage
of Coverlet, if the Coverlet NuGet package is in place.

Also changes the logic to see where OpenCover should be invoked.