(#110) Change logic for SonarQube/DependencyCheck
In a previous commit, a change was made to not allow setting of whether
to run SonarQube and DependencyCheck from the recipe.cake or the command
line. However, based on a follow up discussion with the Operations team,
we have decided that we need this flexibility.  This will allow us to
control exactly when/where SonarQube and DependencyCheck are run.

Both of these will be false by default, and only set within the required
CI builds when needed.