(#129) Remove interaction between Tasks
In the original version of Cake.Recipe, the closing/maintenance of the
release is performed when the tag is placed onto the repository.  This
is no longer the case, as there is a slightly different process that we
follow.

This commit removes the requirement for the publish task to run on a non
local build, etc. as this will likely now be run locally on a developer
VM, until it is integrated fully into a release process.

In addition, remove the publish task from the top level CI task, as it
will now only be run locally, for now.