(#129) Adjust tasks to know when running on GitLab
This commit adds in a new parameter, RepositoryHostedInGitLab, which is
used to determine whether additonal arguments are passed into
GitReleaseManager to trigger execution on GitLab.

This will need to be set in the recipe.cake file for repositories that
want to make use of it. In future, it would be great if this can be
autodetected, either in Chocolatey.Cake.Recipe, or within
GitReleaseManager itself, but for now, this is an ok requirement.

In addition, an extra call to grm publish is added when running in
GitLab, since on GitHub this step is done through the GitHub UI, so it
is only required when running in GitLab.