(#129) Change credential name
Now that GitReleaseManager supports both GitHub and GitLab, having a
Token called GitHub being used, doesn't make sense anymore.

This commit creates a new standalone parameter which is set via a new
Environment Variable called GITRELEASEMANAGER_PAT, which can be set as
required to be either a Token for GitHub or GitLab.