(#139) Remove usage of SetFormatOutput
This is causing a build failure when running on TeamCity. This is a
known issue as per:

https://github.com/SharpeRAD/Cake.Powershell/issues/19

Which is worked around by removing the call to SetFormatOutput. We
weren't able to verify this when we were running the build directly
while RDP'ed into the TeamCity Build Agent, but going to take this
hotfix version of Chocolatey.Cake.Recipe for a spin to verify.