(#139) Reset everything to default values
There is a suggestion on this blog post:

https://geoffhudik.com/tech/2017/11/17/test-coverage-with-opencover-xunit-cake-and-teamcity/

That it is necessary to:

> Okay, OutputToAppConsole has to be explicitly set to false when
running non-interactively.

However, this isn't required on the other usages of StartPowerShellFile
that is in use elsewhere in the recipe.

To rule out other things, let's reset everything back to the default
when calling new PowerShellSettings.  If there is still a problem, we
can do another try with OutputToAppConsole setting to false.