(#130) Install PowerShell Module to Current User
Update the code that installs the PSScriptAnalyzer module to ensure the
NuGet Package Provider is installed and correctly loaded for the
CurrentUser. This ensures the user running the build does not need to be
running as an administrator, nor do they need to already have the NuGet
Package Provider installed and up to date.