(#144) Ensure artifact upload is the last action
WIthin the CopyBuildOutput method, when it is a Web Application that is
being built, the MSBuild.log file is used again to capture some
additional output.  This means that there will sometimes be a change to
the file, and therefore the artitfact upload can't be completed.

By moving the upload to be the last thing that is done in the task, it
should be able to complete successfully.