(#145) Enable skipping of Yarn analysis
During the Dependency-Check task, the Yarn Audit analysis can be
triggered even when it is not needed. This results in the task, and
build, failing.

By adding a new tool setting, DependencyCheckDisableYarnAudit,
the Yarn Audit analysis can be explicitly disabled when required.
This new parameter defaults to `false`.

This change should be revisited when Cake.DependencyCheck and
DependencyCheckTool are updated.