(#152) Only run signing steps on Windows
All of the signing steps in Chocolatey.Cake.Recipe require one of two
things. Either:

- usage of Windows PowerShell
- usage of signtool.exe (with a hard coded path in Windows)

Neither of these things are found on a Linux/Mac build runner.  As such
we need to add criteria to only run tasks on Windows.

This criteria wasn't "needed" previously, since there was another
criteria present to only run when on TeamCity, or the signing
certificate was in place. When the new verify task, which doesn't have
this same criteria, was added, it started failing the build, as the
other tasks weren't being explicit about this requirement.

This commit addresses this underlying problem, by adding an explicit
criteria to each step related to signing to ensure running on Windows
so that it isn't forgotten about in the future.