(#153) Skip signing verification on PR build
The new verify signing task was running on PR builds, which was causing
the PR build to fail when there were signatures that were invalid.

This stops the flow of doing additional testing on the PR builds, as the
arfifacts would not have been created.

We should skip the verification when running on a PR build (or when on a
non-releaseable branch), that way, the additional task of performing
the re-signing can be completed as a follow up piece of work.