(#154) Verify PowerShell scripts on develop branch
In the previous release, a change was made to try to control "when" the
verification of PowerShell scripts would be completed.  This change
neglected to allow running of this task on the develop branch, and
instead only ran on "releaseable" branches, namely master, hotfix and
release.  To allow catching of PowerShell scripts that haven't been
signed, we need to run the verification task on all "main" branches, i.e.
master, develop, release, hotfix and support.  In other words, any
branches that a PR can be merged into.

This commit fixes the criteria for this to be in place, and also adds
the concept of Support branches to the BranchType enumeration, which
wasn't included previously.