(#155) Create zip archive for signed scripts
This updates the signing process of PowerShell scripts, to instead of
uploading them as artifacts in an uncompressed way that requires
downloading multiple files, we instead create a zip archive of the
scripts while preserving the directory structure where the files came
from.

This allows it to be easier in the future when changes have been made,
and we only need to extract a single archive without having to manually
figure out where the files need to be located.