(#162) Rework prerelease substring
This reworks the way we handle the prerelease version. Before if you had
a less than 9 character branch that started with a number, we would
prepend the letter a, then take the first 9 characters. This would
result in an exception. Further, if you had a greater than 10 character
branch name, we would remove the `-` in it and take the first 10
characters from that. With an 11 character branch with 2 `-` this would
be an error as the max length to take is 9. Instead now we will work
just on the part of the string we're fixing, with the max length being
the last one handled.