(GH-171) Enable Dependency-Check central DB
Adds parameters related to configuring Dependency-Check to use a central database
to cache all vulnerability data rather than needing to download it fresh on every
execution.
Database credentials, including the connection string, username, and password are
bring created as a credential object.
Related to this change, an API key is now recommended when downloading vulnerability
data (otherwise heavy rate limits are imposed) and so a parameter for this has also
been added.
The runner tool did not natively support passing through the command line argument
for this API key, which was similar to a previous change for disabling Yarn analysis,
and so this argument is injected into the called batch file when required.