(#136) Allow dry run of task
It is good to be able to run a dry-run version of the notification
sending, just so that we can verify what is going out is what we want.
This commit adds the ability to use --dry-run=true to the command to
"see" the output of what is about to be sent to each notification.