(#164) Use version number in URLs
Previously, the generated URL for release notes used the following
format:

https://docs.chocolatey.org/en-us/choco/release-notes/#july-26-2023

Where the date for the release was included.  Going forward, the
release notes are going to use the version number, similar to the
following:

https://docs.chocolatey.org/en-us/choco/release-notes/#v2.4.2

The commit removes the calculation of the date for the release, and
instead simply uses the version number, prefixed with a "v".