(#132) Fix name of release notes task
Using Publish-GitHub-Release no longer makes sense, now that we have a
GitLab provider.  To make things consistent, this commit renames the
task to be Publish-Release-Notes, which matches with the other tasks
for Export-Release-Notes and Create-Release-Notes.  To provide an
element of backwards compatibility, an alias has been created for
Publish-GitHub-Release, which simply takes a dependency on the new task.