(#188) Don't reset Eazfuscator settings
Previously we reset the Eazfuscator settings on DotNet builds. This
means that the verbosity level selected was not being passed through. As
well, strongname signing keys if present also would not be passed in.
Instead of resetting the settings, we now only reset the setting that we
explicitly touch in the for loop.