(#211) Ensure SolutionVersion.cs file is created
There is a code path which can be executed where GitVersion is not
executed, i.e. when using:

./build.bat --ShouldRunGitVersion=false

When doing this, the code generates a "fake" set of version numbers,
however, those version numbers never made their way into a generated
SolutionVersion.cs file (which is "needed" to have a successful build.

This commit addresses this issue by creating a new method named
GenerateSolutionVersionFile, which is called now from two places.  The
first, where the code was previously, and the second is the if
statements which gates the running of GitVersion.  Now, the "fake"
asserted version numbers are passed into this new method, which then
ensures that the SolutionVersion.cs is created, and the build continues.