(#137 #166) Add prompts for credentials
By default, Chocolatey.Cake.Recipe uses environment variables for
sensitive information that is needed when taking an action. For
example, the token that is used when running GitReleaseManager.  When
these environment variables are not setup, these tasks typically can't
run, and there is no information about what _can_ be done to allow them
to run.

This commit aims to address this by providing information about _what_
environment variable can be set in the future to provide this value,
but also to accept direct input of the values from the command line,
when they are not present.

On each credential class, a new static FetchCredentials method has been
created.  This method will attempt to gather the required information
from the environment variables, and should they not be found, the
Cake.Prompt addin is used, which prompts the user to provide that
information directly at the command line.  Once provided, the Task
continues as per normal.

Any existing HasCredentials, or Can*** methods, have been removed,
since now the necessary information _will_ be provided, either through
the environment variables, or directly through the command prompt.

The introduction of this Prompt on any production builds should be
none, since the required information is already in place, however, when
running locally, when not setup, the prompt will allow the user to be
guided on what information is needed, and also how it can be provided
in the future.