(#282) Find MSBuild tool with an MSBuild.exe
This commit updates the logic for detecting the MSBuild tool. Our builds
encountered this issue: https://github.com/NuGet/Home/issues/14349.

While investigating it, it was found that dnnsoftware encounted a
similar issue and fixed it here:
https://github.com/dnnsoftware/Dnn.Platform/pull/6575. This commit
updates the methods we use to call VSWhere to identify the MSBuild to
return based on actually having an MSBuild.exe file.